migrate

+++ db/migrate/20150725210934_create_rates.rb
@@ -2,7 +2,7 @@ class CreateRates < ActiveRecord::Migration
   def change
     create_table :rates do |t|
       t.string :name
-      t.double :price
+      t.float :price
       t.references :district, index: true, foreign_key: true
       t.references :district, index: true, foreign_key: true
